Disentangled alingnspikes.m
Alignspikes.m no longer saves its outcome. Split up functions to make it less clunky. Idea is to use alignedspikes.m with flexible Srise, depending on which stimulus one wishes to align to. Work still in progress, current version working as it should. Proper integration in rest of code to be done.